rename loggers.py to logs.py, use python logger in _plugins.py
