+++ testing/logging_tests.py
@@ -2,7 +2,7 @@ import logging
 import sys
 import unittest
 
-from nanome._internal.loggers import NTSLoggingHandler
+from nanome._internal.logs import NTSLoggingHandler
 from nanome import Plugin, PluginInstance
 from nanome.util import Logs
 
@@ -34,7 +34,7 @@ class LoggingTestCase(unittest.TestCase):
 
     @patch('nanome._internal._plugin._Plugin._loop')
     @patch('nanome._internal._plugin.Network._NetInstance')
-    @patch('nanome._internal.loggers.NTSLoggingHandler.handle')
+    @patch('nanome._internal.logs.NTSLoggingHandler.handle')
     def test_nts_handler_called(self, handle_mock, netinstance_mock, loop_mock):
         """Assert logs get forwarded to NTS."""
         remote_logging = "True"
@@ -53,7 +53,7 @@ class LoggingTestCase(unittest.TestCase):
 
     @patch('nanome._internal._plugin._Plugin._loop')
     @patch('nanome._internal._plugin.Network._NetInstance')
-    @patch('nanome._internal.loggers.NTSLoggingHandler.handle')
+    @patch('nanome._internal.logs.NTSLoggingHandler.handle')
     def test_nts_handler_not_called(self, handle_mock, netinstance_mock, loop_mock):
         """Assert logs don't get forwarded to NTS if remote-logging is False."""
         remote_logging = False
